FITRUN-132 refactor: 목표 도메인 패키지 분리
