+++ src/main/kotlin/com/yapp/yapp/user/api/request/RunningPurposeRequest.kt
@@ -1,6 +1,6 @@
 package com.yapp.yapp.user.api.request
 
-import com.yapp.yapp.user.domain.RunningPurposeAnswerLabel
+import com.yapp.yapp.user.domain.goal.RunningPurposeAnswerLabel
 
 data class RunningPurposeRequest(
     val runningPurpose: RunningPurposeAnswerLabel,
