+++ src/main/kotlin/com/yapp/yapp/user/domain/goal/UserGoalRepository.kt
@@ -1,5 +1,6 @@
-package com.yapp.yapp.user.domain
+package com.yapp.yapp.user.domain.goal
 
+import com.yapp.yapp.user.domain.User
 import org.springframework.data.repository.CrudRepository
 
 interface UserGoalRepository : CrudRepository<UserGoal, Long> {
