+++ src/main/kotlin/com/yapp/yapp/user/domain/onboarding/OnboardingQuestionType.kt
@@ -24,7 +24,4 @@ enum class OnboardingQuestionType {
 
     /** 페이스 인지 */
     PACE_AWARENESS,
-
-    /** 목표 */
-    GOAL,
 }
