+++ src/test/kotlin/com/yapp/yapp/document/user/UserDocumentTest.kt
@@ -94,7 +94,7 @@ class UserDocumentTest : BaseDocumentTest() {
                 .requestBodyField(
                     fieldWithPath("answers").description("온보딩 설문조사 리스트"),
                     fieldWithPath("answers[].questionType").description("질문 타입(Enum)"),
-                    fieldWithPath("answers[].answer").description("답변 타입(Enum). 답변에 따라 A, B, C, D 중 하나입니다."),
+                    fieldWithPath("answers[].answer").description("답변 타입(Enum). 답변에 따라 A, B, C 중 하나입니다."),
                 )
 
         val restDocsResponse =
@@ -142,7 +142,7 @@ class UserDocumentTest : BaseDocumentTest() {
                     fieldWithPath("result.answerList[].questionType").description("질문 타입(Enum)"),
                     fieldWithPath(
                         "result.answerList[].answer",
-                    ).description("응답 타입(Enum) 답변에 따라 A, B, C, D 중 하나입니다. 러닝 목표만 선택지가 4개이고 나머진 3개입니다."),
+                    ).description("응답 타입(Enum) 답변에 따라 A, B, C 중 하나입니다."),
                 )
 
         val restDocsFilter =
@@ -185,7 +185,7 @@ class UserDocumentTest : BaseDocumentTest() {
                 .requestBodyField(
                     fieldWithPath("answers").description("수정할 온보딩 설문조사 리스트"),
                     fieldWithPath("answers[].questionType").description("질문 타입(Enum)"),
-                    fieldWithPath("answers[].answer").description("답변 타입(Enum). 답변에 따라 A, B, C, D 중 하나입니다."),
+                    fieldWithPath("answers[].answer").description("답변 타입(Enum). 답변에 따라 A, B, C 중 하나입니다."),
                 )
 
         val restDocsResponse =
@@ -196,7 +196,7 @@ class UserDocumentTest : BaseDocumentTest() {
                     fieldWithPath("result.answerList[].questionType").description("질문 타입(Enum)"),
                     fieldWithPath(
                         "result.answerList[].answer",
-                    ).description("답변 타입(Enum). 답변에 따라 A, B, C, D 중 하나입니다. 러닝 목표만 선택지가 4개이고 나머진 3개입니다."),
+                    ).description("답변 타입(Enum). 답변에 따라 A, B, C 중 하나입니다."),
                 )
 
         val restDocsFilter =
@@ -222,7 +222,7 @@ class UserDocumentTest : BaseDocumentTest() {
             RequestFixture.onboardingRequest(
                 answers =
                     listOf(
-                        OnboardingAnswerDto(OnboardingQuestionType.GOAL, OnboardingAnswerLabel.D),
+                        OnboardingAnswerDto(OnboardingQuestionType.RUNNING_EXPERIENCE, OnboardingAnswerLabel.C),
                     ),
             )
 
@@ -236,51 +236,4 @@ class UserDocumentTest : BaseDocumentTest() {
             .then()
             .statusCode(200)
     }
-
-//    @Test
-//    fun `러닝 목표 조회 API`() {
-//        // given
-//        val restDocsRequest =
-//            request()
-//                .requestHeader(
-//                    headerWithName("Authorization").description("엑세스 토큰 (Bearer)"),
-//                )
-//
-//        val restDocsResponse =
-//            response()
-//                .responseBodyFieldWithResult(
-//                    fieldWithPath("result.questionType").description("질문 타입(Enum)"),
-//                    fieldWithPath(
-//                        "result.answer",
-//                    ).description("응답 타입(Enum) A(다이어트), B(건강 관리), C(체력 증진), D(대회 준비) 중 하나입니다."),
-//                )
-//
-//        val restDocsFilter =
-//            filter("사용자 API", "목표 조회")
-//                .tag(Tag.USER_API)
-//                .summary("목표 조회 API")
-//                .description("러닝 목표를 조회하는 API 입니다.")
-//                .request(restDocsRequest)
-//                .response(restDocsResponse)
-//                .build()
-//        val accessToken = getAccessToken()
-//        val request = RequestFixture.onboardingRequest()
-//        RestAssured.given()
-//            .body(request)
-//            .header(HttpHeaders.CONTENT_TYPE, "application/json")
-//            .header("Authorization", "$accessToken")
-//            .`when`().post("/api/v1/users/onboarding")
-//            .then()
-//            .statusCode(201)
-//        // when
-//
-//        // then
-//        RestAssured.given(spec)
-//            .filter(restDocsFilter)
-//            .header(HttpHeaders.CONTENT_TYPE, "application/json")
-//            .header("Authorization", "$accessToken")
-//            .`when`().get("/api/v1/users/onboarding/goal")
-//            .then()
-//            .statusCode(200)
-//    }
 }
